+++ R/zzz.R
@@ -1,5 +1,3 @@
 #' @import rlang
+#' @import tidyselect
 NULL
-
-# https://github.com/r-lib/tidyselect/issues/201
-utils::globalVariables("where")